refactor(gui): break up netwatch gui
