+++ netwatch/ui/__init__.py
@@ -1,0 +1,2 @@
+from netwatch.ui.gui import GUI
+from netwatch.ui.config import config_window
